Auth fix v2
