+++ .env.example
@@ -1,0 +1,4 @@
+# Add your environment variables here
+# VITE_API_URL=
+# VITE_SUPABASE_URL=
+# VITE_SUPABASE_ANON_KEY=
