+++ src/components/language-toggle.tsx
@@ -1,17 +1,22 @@
-import { useI18n } from "@/lib/i18n";
+import { useI18n } from "@/lib/i18n-hooks";
 
 export function LanguageToggle() {
   const { locale, setLocale } = useI18n();
+
+  const toggleLanguage = () => {
+    setLocale(locale === "en" ? "hi" : "en");
+  };
+
   return (
-    <select
-      aria-label="Language"
-      className="bg-transparent border rounded-md px-2 py-1 text-sm"
-      value={locale}
-      onChange={(e) => setLocale(e.target.value)}
+    <button
+      onClick={toggleLanguage}
+      className="flex items-center gap-2 px-3 py-2 text-sm font-medium text-muted-foreground hover:text-foreground transition-colors"
     >
-      <option value="en">EN</option>
-      <option value="hi">HI</option>
-    </select>
+      <span className="text-lg">
+        {locale === "en" ? "🇮🇳" : "🇺🇸"}
+      </span>
+      <span>{locale === "en" ? "हिंदी" : "English"}</span>
+    </button>
   );
 }
