+++ src/contexts/AuthContext.ts
@@ -1,0 +1,2 @@
+export { AuthContext, type AuthContextType } from './auth.types';
+export { AuthProvider } from './AuthProvider';
