+++ src/contexts/index.ts
@@ -1,0 +1,4 @@
+// Re-export all authentication-related components and types
+export { AuthProvider } from './AuthProvider';
+export { AuthContext, type AuthContextType } from './auth.types';
+export { useAuth } from '@/hooks/useAuth';
